#34 allow containers inside containers in UI
